make registrationManager use uint64s

+++ keeper/registration_manager_test.go
@@ -10,7 +10,7 @@ import (
 
 var address = common.HexToAddress("0x0000000000000000000000000000000000000123")
 var checkGasLimit = int64(10_000)
-var cooldown = uint(3)
+var cooldown = uint64(3)
 
 func setupRegistrationManager(t *testing.T) (*store.Client, RegistrationManager, func()) {
 	db, cleanup := store.SetupTestDB(t)
